(chore) move Review subcomponent to subfolder
